Add Game state
move status from Board
