+++ src/tic-tac-toe/Board.tsx
@@ -1,46 +1,23 @@
 import * as React from "react";
 import "./Board.css";
-import { Square, Value } from "./Square";
-
-export type OnClick = () => void;
-
-// prettier-ignore
-type Squares = [
-  Value, Value, Value,
-  Value, Value, Value,
-  Value, Value, Value
-  ];
+import { OnClick, Squares } from "./Game";
+import { Square } from "./Square";
 
 interface IBoard {
   squares: Squares;
+  handleClick: (i: number) => OnClick;
 }
 
-export class Board extends React.Component<any, IBoard> {
-  constructor(props: Readonly<any>) {
-    super(props);
-    this.state = {
-      squares: Array(9).fill(null) as Squares
-    };
-  }
-
+export class Board extends React.Component<IBoard> {
   render() {
-    const status = "Next player: X";
-
     return (
       <div>
-        <div className="status">{status}</div>
         <div className="board center-block">
-          {this.state.squares.map((v, i) => (
-            <Square value={v} onClick={this.handleClick(i)} />
+          {this.props.squares.map((v, i) => (
+            <Square value={v} onClick={this.props.handleClick(i)} />
           ))}
         </div>
       </div>
     );
   }
-
-  protected handleClick: (i: number) => OnClick = i => () => {
-    const squares = [...this.state.squares] as Squares; // copy
-    squares[i] = "X";
-    this.setState({ squares });
-  };
 }
